Refactor and rename variables
