+++ apps/contentful-apps/components/content-import/index.ts
@@ -3,4 +3,4 @@ export * from './FileInput'
 export * from './ReferenceFieldMapping'
 export * from './TagSelect'
 export * from './FileDataTable'
-export * from './FieldMapping'
+export * from './PrimitiveFieldMapping'
