+++ apps/contentful-apps/constants/index.ts
@@ -4,3 +4,6 @@ export const MIDEIND_TRANSLATION_API_BASE_URL =
   'https://stafraentisland.greynir.is/translate'
 export const DEFAULT_LOCALE = 'is-IS'
 export const DEV_WEB_BASE_URL = 'https://beta.dev01.devland.is'
+
+export const TITLE_SEARCH = '--title-search'
+export const SLUGIFIED = '--slugified'
